trueval_fullstack_bundle/flask-backend/requirements.txt
flask
airtable-python-wrapper
